+++ lib/qx/draw/vega_lite.ex
@@ -68,7 +68,6 @@ defmodule Qx.Draw.VegaLite do
     end
   end
 
-  defp counts_key_to_label(key) when is_list(key), do: Enum.join(key, "")
   defp counts_key_to_label(key) when is_binary(key), do: key
 
   @doc """